backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
qrcode[pil]==7.4.2
httpx>=0.24.0,<0.25.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-api-python-client==2.136.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
